move intersection tests to set-utils.spec.ts
