+++ editor/src/core/shared/math-utils.spec.ts
@@ -1,5 +1,4 @@
-import { intersection, wrapValue } from './math-utils'
-import * as fc from 'fast-check'
+import { wrapValue } from './math-utils'
 
 describe('math utils', () => {
   describe('wrapValue', () => {
@@ -20,37 +19,6 @@ describe('math utils', () => {
       expect(wrapValue(-4, -7, -3)).toEqual(-4)
       expect(wrapValue(-11, -7, -3)).toEqual(-6)
       expect(wrapValue(-2, -7, -3)).toEqual(-7)
-    })
-  })
-
-  describe('intersection', () => {
-    it('intersection of overlapping sets', () => {
-      const result = intersection([new Set([1, 2, 3, 4]), new Set([3, 4, 5, 6])])
-      expect([...result]).toEqual([3, 4])
-    })
-
-    it('intersection of disjunct sets', () => {
-      const result = intersection([new Set([1, 2, 3]), new Set([4, 5, 6])])
-      expect([...result]).toEqual([])
-    })
-
-    it('intersection is commutative', () => {
-      fc.assert(
-        fc.property(fc.set(fc.nat()), fc.array(fc.nat()), (left, right) => {
-          const resultA = intersection([new Set(left), new Set(right)])
-          const resultB = intersection([new Set(right), new Set(left)])
-          expect([...resultA]).toEqual([...resultB])
-        }),
-      )
-    })
-
-    it('intersection is of a set with itself is the original set', () => {
-      fc.assert(
-        fc.property(fc.set(fc.nat()), (xs) => {
-          const resultA = intersection([new Set(xs), new Set(xs)])
-          expect([...resultA]).toEqual([...xs])
-        }),
-      )
     })
   })
 })
